🔨 refactor(dreams): Refactor dream routes to use dream controller
Consolidated dream routes by utilizing a dedicated dream controller. This change enhances code modularity and maintainability, and centralizes business logic, reducing redundancy across routes.
